+++ khoros/studio/__init__.py
@@ -1,0 +1,14 @@
+# -*- coding: utf-8 -*-
+"""
+:Module:            khoros.studio
+:Synopsis:          This module performs tasks that relate to the Lithium SDK and the Khoros Community Studio Plugin.
+:Usage:             ``from khoros import studio``
+:Created By:        Jeff Shurtliff
+:Last Modified:     Jeff Shurtliff
+:Modified Date:     20 May 2020
+"""
+
+__all__ = ['base']
+
+# Import all submodules by default
+from . import base